requirements.txt
numpy==2.1.1
pywebview==5.2
beautifulsoup4==4.12.3